Fix ctypes ABI mismatch in WaveletFilter's length argument
filter_signal_wt's C signature takes the signal length as int64_t, but
the ctypes argtypes declared it as c_int (32-bit), an ABI mismatch that
risked memory corruption or a wrong buffer length on large arrays.

+++ poriscope/plugins/filters/WaveletFilter.py
@@ -182,7 +182,7 @@ class WaveletFilter(MetaFilter):
         self.fun.restype = None
         self.fun.argtypes = [
             ndpointer(ctypes.c_double, flags="C_CONTIGUOUS"),
-            ctypes.c_int,
+            ctypes.c_int64,
             ctypes.c_char_p,
         ]
